// gql/apiTypes.ts
export type GQLtimestamp = any
// ...
export type GQLjsonb = any
// ...
export interface GQLdeployment_versions {
  build_host_name?: string
  built_at?: GQLtimestamp
  configuration?: GQLjsonb
  deployed_at?: GQLtimestamp

  /**
   * An object relationship
   */
  deployment?: GQLdeployments
  deployment_id?: string
  docker_image?: string
  docker_image_tag?: string
  env?: string
  git_branch?: string
  git_commit?: string
  git_hash?: string
  git_url?: string
  id: string
  kubernetes_deployment_files?: GQLjsonb
  last_commits?: string
  version?: string
}

/**
 * columns and relationships of "deployments"
 */
export interface GQLdeployments {
  db_migration_image?: string
  deployer_role: string
  deployment_type: string

  /**
   * An array relationship
   */
  deployment_versions: GQLdeployment_versions[]

  /**
   * An aggregated array relationship
   */
  deployment_versions_aggregate: GQLdeployment_versions_aggregate
  description: string
  display_name: string
  env?: string
  hyperlinks?: GQLjsonb
  id: string
  last_deployment_timestamp: GQLtimestamp
}

/**
 * aggregated selection of "deployment_versions"
 */
export interface GQLdeployment_versions_aggregate {
  aggregate?: GQLdeployment_versions_aggregate_fields
  nodes: GQLdeployment_versions[]
}

/**
 * aggregate fields of "deployment_versions"
 */
export interface GQLdeployment_versions_aggregate_fields {
  count?: number
  max?: GQLdeployment_versions_max_fields
  min?: GQLdeployment_versions_min_fields
}